moved bitvect to xoxo
