+++ examples/xoxo/xoxo.py
@@ -1,117 +1,131 @@
-from pysmt.shortcuts import FreshSymbol, Int, Or, And, Equals, Plus, Solver
-from pysmt.typing import INT
+from pysmt.shortcuts import FreshSymbol, BV, Or, And, Equals, Plus, Solver, Int
+from pysmt.typing import BVType
 from enum import Enum
 import logging
+
 logger = logging.getLogger('xoxo')
 logger.setLevel(logging.DEBUG)
 ch = logging.StreamHandler()
 
-if __debug__: # start with -O
-    log_level = logging.DEBUG
-else:
+if __debug__:
+    print("start with python -O for debugging")
     log_level = logging.INFO
+else:
+    print("verbose logging on")
+    log_level = logging.DEBUG
+
 ch.setLevel(log_level)
 
 logger.addHandler(ch)
 
+VECT_WIDTH = 4
+
 class Cell(Enum):
-    x = 1
-    o = 10
-    s = 0
+    s = BV(0,VECT_WIDTH) # space
+    x = BV(1,VECT_WIDTH) # x - human player goes first
+    o = BV(2,VECT_WIDTH) # o - cpu player
 
-turns = 0
-sq_size = 3
-test = 'tests/test1.txt'
-test = 'tests/blank.txt'
+x_turns = 0
+o_turns = 0
+x_val = Cell.x.value.constant_value()
+o_val = Cell.o.value.constant_value()
 
-# create symbols for the board
-board = [[FreshSymbol(INT) for _ in xrange(sq_size)]
-            for _ in xrange(sq_size)]
+board = [[FreshSymbol(BVType(VECT_WIDTH)) for _ in range(3)]
+            for _ in range(3)]
 
 solver = Solver()
 
-# setup assertions for board
+# initialise board cells, each one has to be blank, x or o
 for row in board:
     for cell in row:
-        solver.add_assertion(Or([Equals(cell, Int(i.value))
+        solver.add_assertion(Or([Equals(cell, i.value)
              for i in Cell]))
 
-display_board = [[Cell.s for _ in xrange(sq_size)] for _ in xrange(sq_size)]
-
 # load board
+test = 'tests/blank.txt'
 with open(test) as fh:
     for row, line in enumerate(fh.readlines()):
         for col, cell in enumerate(line.strip().split(' ')):
             if cell == Cell.x.name:
-                solver.add_assertion(Equals(board[row][col], Int(Cell.x.value)))
-                display_board[row][col] = Cell.x
-                turns += 1
+                solver.add_assertion(Equals(board[row][col], Cell.x.value))
             elif cell == Cell.o.name:
-                solver.add_assertion(Equals(board[row][col], Int(Cell.o.value)))
-                display_board[row][col] = Cell.o
-
-def get_win_assertion(player):
-    return [
-        # rows
-        Equals(Plus(board[0]), Int(player * sq_size)), 
-        Equals(Plus(board[1]), Int(player * sq_size)), 
-        Equals(Plus(board[2]), Int(player * sq_size)),
-        # cols
-        Equals(Plus([row[0] for row in board]), Int(player * sq_size)),
-        Equals(Plus([row[1] for row in board]), Int(player * sq_size)),
-        Equals(Plus([row[2] for row in board]), Int(player * sq_size)),
-        # diags
-        Equals(Plus([board[0][0], board[1][1], board[2][2]]), Int(player * sq_size)),
-        Equals(Plus([board[0][2], board[1][1], board[2][0]]), Int(player * sq_size))
-        ]
-
-# assertions for players to win
-o_win_assertions = get_win_assertion(Cell.o.value)
-x_win_assertions = get_win_assertion(Cell.x.value)
-
-logger.debug(solver.assertions)
-
-def print_board():
-    for row in display_board:
-        logger.info(" ".join([cell.name.replace('s','.') for cell in row]))
-
-def print_solver_board():
-    logger.debug("solver board")
-    for row in board:
-        logger.debug(" ".join([Cell(int(solver.get_py_value(cell))).name.replace('s','.') for cell in row]))
-
-def get_a_move(player):
-    logger.debug("looking for a move for %s" % player)
-    # prioritise center position
-    if display_board[1][1] == Cell.s and Cell(int(solver.get_py_value(board[1][1]))) == player:
-        return(1,1)
-    for r, row in enumerate(board):
-        for c, cell in enumerate(row):
-            if display_board[r][c] == Cell.s and Cell(int(solver.get_py_value(cell))) == player:
-                return(r,c)
-
-def convert_num_to_indices(num):
-    row = int(num/sq_size)
-    col = num % sq_size
-    return(row,col)
+                solver.add_assertion(Equals(board[row][col], Cell.o.value))
 
 def already_played(row, col):
-    if display_board[row][col] == Cell.s:
+    if solver.get_value(board[row][col]) == Cell.s.value:
         return False
     return True
 
-def play_move(player, row, col):
-    display_board[row][col] = player
-    solver.add_assertion(Equals(board[row][col], Int(player.value)))
+def print_board():
+    for row in board:
+        line = ""
+        for cell in row:
+            if solver.get_value(cell) == Cell.x.value:
+                line += "x"
+            elif solver.get_value(cell) == Cell.o.value:
+                line += "o"
+            else:
+                line += "-"
+            line += " "
+        print(line)
+
+def play_move(p, row, col):
+    logger.debug("adding assertion for player %s at (%d, %d)" % (p.name, row, col))
+    solver.add_assertion(And(Equals(board[row][col], p.value)))
+
+def get_win_assertion(p):
+    return [
+           # rows
+           And(Equals(board[0][0], p.value), Equals(board[0][1], p.value), Equals(board[0][2], p.value)),
+           And(Equals(board[1][0], p.value), Equals(board[1][1], p.value), Equals(board[1][2], p.value)),
+           And(Equals(board[2][0], p.value), Equals(board[2][1], p.value), Equals(board[2][2], p.value)),
+
+           # cols
+           And(Equals(board[0][0], p.value), Equals(board[1][0], p.value), Equals(board[2][0], p.value)),
+           And(Equals(board[0][1], p.value), Equals(board[1][1], p.value), Equals(board[2][1], p.value)),
+           And(Equals(board[0][2], p.value), Equals(board[1][2], p.value), Equals(board[2][2], p.value)),
+
+           # diags
+           And(Equals(board[0][0], p.value), Equals(board[1][1], p.value), Equals(board[2][2], p.value)),
+           And(Equals(board[2][0], p.value), Equals(board[1][1], p.value), Equals(board[0][2], p.value)),
+           ]
+
+# only return options not already played
+def find_all_moves(p):
+    logger.debug("finding all possible moves for %s" % p.name)
+    options = []
+    for r, row in enumerate(board):
+        for c, cell in enumerate(row):
+            if not Equals(board[r][c], p.value) in solver.assertions: # if not already played
+                options.append(Equals(cell, p.value))
+                logger.debug("%d,%d" % (r,c))
+    return options
+
+def pick_new_move(p):
+    logger.debug("picking a move for %s" % p.name)
+    for r, row in enumerate(board):
+        for c, cell in enumerate(row):
+            if not Equals(board[r][c], p.value) in solver.assertions: # if not already played
+                if solver.get_value(cell) == p.value: # and is in the solution
+                    return(r,c)
+
+# used to determine how many counters can be on the board given the current turn
+def get_board_sum():
+    return board[0][0] + board[0][1] + board[0][2] + board[1][0] + board[1][1] + board[1][2] + \
+            board[2][0] + board[2][1] + board[2][2]
+
+def convert_num_to_indices(num):
+    row = int(num/3)
+    col = num % 3
+    return(row,col)
 
 while True:
-
     # get user input and handle errors
     logger.info("-" * 40)
-    logger.info("turn %d" % turns)
+    solver.solve([Equals(get_board_sum(), BV(x_turns * x_val + o_turns * o_val, VECT_WIDTH))])
     print_board()
     try:
-        next_cell = int(raw_input("type a cell (1-9):")) - 1
+        next_cell = int(input("type a cell (1-9):")) - 1
     except ValueError:
         continue
     if next_cell < 0 or next_cell > 8:
@@ -121,70 +135,54 @@ while True:
     row, col = convert_num_to_indices(next_cell)
     if(not already_played(row, col)):
         play_move(Cell.x, row, col)
+        x_turns += 1
     else:
         logger.info("that cell is already taken")
         continue
 
-    turns += 1
-
-    # assertions for next move
-    x_next_move_assertion = Equals(Int((turns+1)*Cell.x.value + (turns*Cell.o.value)), Plus(Plus(board[0]), Plus(board[1]), Plus(board[2])))
-    x_this_move_assertion = Equals(Int((turns-1)*(Cell.x.value + Cell.o.value)+Cell.x.value), Plus(Plus(board[0]), Plus(board[1]), Plus(board[2])))
-    this_move_assertion = Equals(Int((turns)*(Cell.x.value + Cell.o.value)), Plus(Plus(board[0]), Plus(board[1]), Plus(board[2])))
-    o_win_this_turn_assertions = And(Or(o_win_assertions), this_move_assertion)
-    x_win_next_turn_assertions = And(Or(x_win_assertions), x_next_move_assertion)
-    x_win_this_turn_assertions = And(Or(x_win_assertions), x_this_move_assertion)
-
-    # detect x has won
-    logger.debug("checking if x has won")
-    logger.debug(x_win_this_turn_assertions)
-    if solver.solve([x_win_this_turn_assertions]):
-        print_solver_board()
+    # check for x to win
+    if solver.solve([
+                        Or(get_win_assertion(Cell.x)),
+                        Equals(get_board_sum(), BV(x_turns * x_val + o_turns * o_val, VECT_WIDTH))]):
         print("x wins")
-        exit()
+        print_board()
+        exit(0)
+    elif x_turns == 5:
+        print("no win possible")
+        exit(0)
 
-    # can I win next turn?
-    logger.debug("look for o to win")
-    logger.debug(o_win_this_turn_assertions)
-    if solver.solve([o_win_this_turn_assertions]):
-        logger.debug("o can win like this")
-        print_solver_board()
-        result = get_a_move(Cell.o)
-        if result is not None:
-            play_move(Cell.o, result[0], result[1])
-            print_board()
-            print("o wins")
-            exit()
+    # o's turn played by cpu
+    o_turns += 1
 
-    logger.debug("look for x to win next turn")
-    logger.debug(x_win_next_turn_assertions)
-    if solver.solve([x_win_next_turn_assertions]):
-        # can x win next turn? if so stop
-        logger.debug("x can win like this")
-        print_solver_board()
-        result = get_a_move(Cell.x)
-        if result is not None:
-            play_move(Cell.o, result[0], result[1])
-            continue
+    # see if o can win this turn
+    if solver.solve([
+                        Or(get_win_assertion(Cell.o)),
+                        Equals(get_board_sum(), BV(x_turns * x_val + o_turns * o_val, VECT_WIDTH))]):
+        logger.debug("found a way for o to win")
+        result = pick_new_move(Cell.o)
+        play_move(Cell.o, result[0], result[1])
+        print("o wins")
+        print_board()
+        exit(0)
 
-    # otherwise try and win
-    logger.debug("looking for a win")
-    played = False
-    for future_turn in range(turns, 5):
-        move_assertion = Equals(Int((future_turn)*(Cell.x.value + Cell.o.value)), Plus(Plus(board[0]), Plus(board[1]), Plus(board[2])))
-        for assertion in o_win_assertions:
-            logger.debug("trying %s turn %d" % (assertion, future_turn))
-            res = solver.solve([assertion, move_assertion])
-            if res:
-                logger.debug("found a win on turn %d" % future_turn)
-                print_solver_board()
-                row, col = get_a_move(Cell.o)
-                play_move(Cell.o, row, col)
-                played = True
-                break
+    # try to block x next turn (x_turns+1) after both players have played again
+    elif solver.solve([
+                        Or(get_win_assertion(Cell.x)),
+                        And(Or(find_all_moves(Cell.o)), Or(find_all_moves(Cell.x))),
+                        Equals(get_board_sum(), BV((x_turns+1) * x_val + o_turns * o_val, VECT_WIDTH))]):
+        logger.debug("found a way to block x winning next time with board val %d" % ((x_turns+1) * x_val + o_turns * o_val))
+        print_board()
+        result = pick_new_move(Cell.x) # get the winning move for x and play for o
+        play_move(Cell.o, result[0], result[1])
 
-        if played:
-            break
-
-    if not played:
-        exit("I give up")
+    # force solver to find a next move for o, rather than 2 moves for x by specifying possible next moves with find_all_moves()
+    elif solver.solve([
+                        Or(find_all_moves(Cell.o)),
+                        Equals(get_board_sum(), BV(x_turns * x_val + o_turns * o_val, VECT_WIDTH))]):
+        result = pick_new_move(Cell.o)
+        play_move(Cell.o, result[0], result[1])
+    
+    # o can't play
+    else:
+        print("o can't play")
+        exit(0)
